requirements.txt
agora-token-builder==1.0.0
asgiref==3.5.0
Django==4.0.4
django-ckeditor==6.4.2
django-countries==7.3.2
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.1.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
